correção de erros

+++ perfil/views.py
@@ -77,16 +77,17 @@ def editar_informacoes(request):
             user = Usuario.objects.get(username=request.user.username)
             
             # Remove foto antiga
-            caminho_foto_antiga = user.foto.path
-            if os.path.exists(caminho_foto_antiga):
-                os.remove(caminho_foto_antiga)
+            if request.FILES.get('input-file'): # verifica se há arquivo
+                user.foto = list_data[4]
+                caminho_foto_antiga = user.foto.path
+                if os.path.exists(caminho_foto_antiga):
+                    os.remove(caminho_foto_antiga)
             
             # Atualiza dados
             user.first_name = list_data[0]
             user.last_name = list_data[1]
             user.email = list_data[2]
             user.telefone = list_data[3]
-            user.foto = list_data[4]
             
             user.save()
